refactor: #625 relocate files
Simply put files in better places so that they can be found and reused.
