+++ packages/liferay-npm-bundler/src/steps/adapt/angular-cli/index.ts
@@ -5,6 +5,7 @@
 
 import {project} from '../../../globals';
 import * as log from '../../../log';
+import adaptStaticURLsAtRuntime from '../../../transform/js/operation/adaptStaticURLsAtRuntime';
 import {runPkgJsonScript} from '../../../util/run';
 import {
 	WebpackBundles,
@@ -14,7 +15,6 @@ import {
 	processPackageJson,
 	processWebpackBundles,
 } from '../index';
-import adaptStaticURLsAtRuntime from '../transform/js/operation/adaptStaticURLsAtRuntime';
 import tweakAttachmentToDOM from './tweakAttachmentToDOM';
 
 const {pkgJson} = project;
