+++ packages/liferay-npm-bundler/src/steps/adapt/bundler-project/replaceWebpackJsonp.ts
