+++ packages/liferay-npm-bundler/src/steps/adapt/index.ts
@@ -18,11 +18,11 @@ import path from 'path';
 
 import {bundlerGeneratedDir, project} from '../../globals';
 import * as log from '../../log';
+import exportModuleAsFunction from '../../transform/js/operation/exportModuleAsFunction';
+import namespaceWepbackJsonp from '../../transform/js/operation/namespaceWepbackJsonp';
+import replace from '../../transform/text/operation/replace';
 import {copyFiles, findFiles} from '../../util/files';
 import Renderer from '../../util/renderer';
-import exportModuleAsFunction from './transform/js/operation/exportModuleAsFunction';
-import namespaceWepbackJsonp from './transform/js/operation/namespaceWepbackJsonp';
-import replace from './transform/text/operation/replace';
 
 /**
  * Description of framework's webpack build output so that adapted modules can
