+++ packages/liferay-npm-bundler/src/steps/adapt/templates/export-module.ejs
