+++ packages/liferay-npm-bundler/src/steps/adapt/templates/manifest-module.ejs
