+++ packages/liferay-npm-bundler/src/steps/webpack/loader/imports-loader.ts
@@ -7,7 +7,7 @@ import {transformJsSource} from 'liferay-js-toolkit-core';
 
 import {project} from '../../../globals';
 import report from '../../../report';
-import transformImports from './transformImports';
+import transformImports from '../../../transform/js/operation/transformImports';
 
 /**
  * Webpack loader to substitute require calls for imported modules. This loader
