+++ packages/liferay-npm-bundler/src/transform/js/operation/exportModuleAsFunction.ts
