+++ packages/liferay-npm-bundler/src/transform/js/operation/transformImports.ts
